Add a cast for Nintendo
They insist on using "enum class" and making their parallel version of
Berkely sockets, and I am 100% confident that literally not one single
person actually wants this.

P4:7324300

+++ src/tier1/netadr.cpp
@@ -431,7 +431,7 @@ bool CIPAddress::SetFromSockadr(const void *addr, size_t addr_size, uint16 *punP
 		Assert( false );
 		return false;
 	}
-	switch ( s->sa_family )
+	switch ( (int)s->sa_family )
 	{
 		case AF_INET:
 		{
